Move ExecuteInternalCommand to private file
